tip(1): Remove the clearing of the schg flag of cu(1) during installation.
A long, long time ago (until 2004), when cu(1) was still a separate
program, its schg flag was set after installation, so it had to be
cleared on the installed cu before installing the new one over it.

In 0f3db4ffe, we removed cu(1) as a separate program and made it a
link to tip(1). In this commit, the setting of schg was commented
out but the clearing part had to be kept (for upgrading).

We can now remove the clearing of the flag as well, because we can
safely assume that since September 2004, all schg flags of cu(1)s
out there have been cleared (i.e., all systems that would ever want
to upgrade to master have a world newer than 0f3db4ffe).