tzsetup: Sync with FreeBSD.
Primarily this is an update for the new libdialog, but while I'm here:

- Add a menu entry for UTC in the main menu.
- Add support for chrooted installs.
- Add examples to the man-page.
- Instead of having to know which timezone was picked last time, you
now can run "tzsetup -r" which will reinstall the last choice. This
data is recorded in /var/db/zoneinfo.
- Add a new options (-s) that, when specified, skips the question about
adjusting the clock to UTC. That avoids to write on /etc/wall_cmos_clock
which is useful in some cases (example: host user in a jail).
- Remove paths.h and embed it in the source.
- Fix style(9) issues.

Obtained-from:   FreeBSD