ccdconfig(8): When ccd can't be kldloaded, explain a possible reason.
If /etc/ccd.conf is present, /etc/rc.d/ccd will run ccdconfig in the
early boot when only root is mounted read-only. On HAMMER systems (or
any other system where /boot isn't part of the root partition) it can
not automatically load ccd.ko at this point. This was different on
UFS where /boot was usually not separate.

So in case it fails, give the user a better message that explains the
options he has.

Dragonfly-bug: <http://bugs.dragonflybsd.org/issues/2439>
Reported-by:   Christer Solskogen