libedit: Update to version 2012-12-13
There have been three releases since libedit was imported as a vendor
branch, but the changelog is simply "sync with NetBSD" for each one.

The main item of interest is that widechar support has been added
for real this time.  Previously all the man pages were in place but
the library didn't actually have the support.

A minor item of note is that NetBSD finally implemented the
rl_completion_word_break_hook which reduces the number of files
with local modifications.