kernel/oce: Sync oce(4) driver with FreeBSD.
An update by Emulex. Most notably, it adds support for Skyhawk adapters.

The port to DragonFly wasn't tested on a Skyhawk but looks like it should
just work[tm].

Thanks to Mihai Carabas for testing it on a BladeEngine 3 to see that I
didn't break anything which previously worked.

Taken-from: FreeBSD