Makefile.inc1: Fix build when MAKEOBJDIRPREFIX is provided
While testing a patch that removed symbols from libc, Joris found his
build would break on leaf with MAKEOBJDIRPREFIX set in the command-line.
While this is a misuse of the feature per bsd.obj.mk comments, the issue
was reproduced on my machine.  The object files from btools, ctools, and
world were getting placed in common directories, so world binaries were
being built with bootstrap and crosstools objects.

I do not know if this has always been the case, or if this started
happening when bmake replaced legacy make.  In any case, changing the
way DESTDIRBASE is defined solved the build issues on the local machine.