Restore ability to build Release 3.4 branch from master
Currently it is not possible to build the Release 3.4 branch with a system
built from the head of the master branch.  The problem lies with libgrep
which attempts to define a symbol that now exists on the host system, so
bootstrapping grep fails.  This small patch will omit the wcwidth.c
source file if the host version is 503 (__DragonFly_version 300503) or
higher, and only when bootstrapping.  This fix is only applied to the 3.4
Release branch as it is not needed on the master branch.