Update pkg bootstraps
The naming scheme for pkg bootstraps is inadequate.  Even if it were to
include the release, library updates could necessitate the generation
of a new bootstrap more than once per release.  Therefore, I've added
a serial in the name.  The name is now:

  bootstrap_pkg_${SERIAL}_x86-${32,64}.tar.xz

Bootstrap 01 has been created and put on mirror master.  The original
bootstraps stopped working after the locales update.