Fix world build bootstrapping issue
The new version of find(1) uses a new locale function called rpmatch.
The use of this function breaks the world build on systems that don't
have the new locale functionality yet, e.g. Release 3.4.

To fix this, I added a BOOTSTRAPPING macro that will check the
first letter of the response character array rather than use the
locale function.  This modification appears only in the bootstrap tool.