libldns, drill(1): Update to version 1.6.16
Many dports that require libldns were not building because the library
was detected, but not the ldns.h header (along with its 30+ friends).
The headers were available in contrib, but not installed.  However, the
version 1.6.11 is too old for at least some of the ports, so it became
necessary to update LDNS to the latest version.