mxge: Implement MSI-X support; multiple RX rings could be enabled
One thing need to note is the interrupt moderation when MSI-X is
enabled.  On the PCIE-8AL-C, it looks like that the interrupt rate
set to the chip means total interrupt rate, NOT per MSI-X vector
interrupt rate: e.g. Given the interrupt rate is set too 8000 and 8
MSI-X vectors are allocated.  If two MSI-X vectors are active, then
the interrupt rate for each MSI-X vector will be ~4000.  If all
MSI-X vectors are active, then the interrupt rate for each MSI-X
vector will be ~1000.  This is kind of interrupt moderation for
MSI-X is very unfriendly ...

MSI-X is not enabled by default yet.  You could set tunable
hw.mxge.num_slices or hw.mxgeX.num_slices to 0 or any value greater
than 1 to enable MSI-X.

RSS key is not properly setup yet.