Remove the <sys/syscall-hide.h> header file.
Nothing includes it, nor are the HIDE_BSD() and HIDE_POSIX() macros
actually defined anywhere.

Also, remove support for the associated POSIX, BSD and NOHIDE keywords
from the syscalls.master files. It makes the patch quite large due to
touching most lines in the syscalls.master files but it results in
zero changes in the generated files.