kernel: Generate miidevs.h, pccarddevs.h and pcidevs.h on the fly.
It removes the need to regenerate those header file after first editing
the associated list of IDs (miidevs, pccarddevs or pcidevs). After this
commit, editing the list alone is enough to add IDs.

We already did it like that for usb4bsd's usbdevs.h before. This commit
adjusts things for the remaining ID lists.