gold 2.24: Sync with binutils 2.22
I had forgotten that the build for gold changed on DragonFly between
binutils 2.21 and binutils 2.22.  When I based 2.24 on 2.21, I lost
these change (namely a common Makefile.gld and the support for the
new utility incremental dump).

I also forgot the update the generated yyscript.[ch] files.
This commit brings binutils 2.24 in-line with 2.22.