netmap: change header includes
Suggested-by: sephe

FWIW, kqfilter() port is not really going well as poll() is doing
a bunch of loosely related drive-by actions that somehow form the
API.  The best approach would be to thin out the implementation to
at least get the bare core functionality working.  That'll also
include writing new userland glue.  Feels more like a fork than a
port that way.  :(