iconv: Sync#2 - Support ICONV_{GET,SET}_ILSEQ_INVALID
Add ICONV_{GET,SET}_ILSEQ_INVALID iconvctl. GNU iconv returns EILSEQ
when there is an invalid character in the output codeset while it is
valid in the input.  However, POSIX requires iconv() to perform an
implementation-defined conversion on the character.  So, Citrus iconv
converts such a character to a special character which means it is
invalid in the output codeset.

This is not a problem in most cases but some software like libxml2
depends on GNU's behavior to determine if a character is output as-is
or another form such as a character entity (&#NNN;).

Taken verbatim from:
	FreeBSD SVN 258537 (24 NOV 2013)