libutil: Remove expand_number(3)
This new function is functionally identical to dehumanize_number(3).
There is no apparently benefit to have two implementations so I am
removing the new one which is not used anywhere.  The manpage was never
hooked into the build (a mistake) so removing it leaves no leftovers.

At the same time, restore the humanize_number man page which had seen
the dehumanize_number function description removed accidently during
the recent libutil sync.

Discussed-with: swildner and dillon