colldef(1)/mklocale(1): Define __heedresult when BOOTSTRAPPING.
These two utilities are bootstrap tools, and due to already existing
surgery take part of their includes from the src/, so as a consequence
we have to define __heedresult too in the BOOTSTRAPPING case.

This is actually the sort of issues which the bootstrap-tools concept
is meant to protect us against. :-(