Stop suppressing -Wuninitialized in userland for WARNS >= 2 and <= 4.
This is from 2001 (via FreeBSD, who still have it). Since a lot of
important stuff is in that WARNS range (libc, most prominently) and
since the option has found some good bugs for us (just see the MFCs
of recent days), we definitely want it to be enabled. None of the
issues it reported turned out to be a false positive.

Do, however, turn off -Wmaybe-uninitialized in that WARNS range for
now.

Also, while here, turn on -Winit-self too.