kernel/acpi: Add quirks framework (and some quirks) from FreeBSD.
I don't know how common the signatures from the supplied quirks
are nowadays (the last addition was in 2005) but the framework
could nevertheless be useful.

For now, the ACPI_Q_MADT_IRQ0 quirk remains unimplemented because
of the way the associated code is implemented in DragonFly. This
will need further work in the future.

In-discussion-with: sephe