kernel/acpi: Fix memory and port I/O functions in the OS services layer.
This makes AcpiOsReadPort(), AcpiOsWritePort(), AcpiOsReadMemory()
and AcpiOsWriteMemory() match the behavior requested in the "ACPI
User Guide and Programmer Reference".

When the width of the location that is read from or written to is less
than the (return) value's width, the read functions must ensure the
rest of the bits are zero and the write functions must write no more
than the location's width.

This should fix at least a number of temperature reading issues we
were seeing, such as the messages we were getting lately:

acpi_tz0: _TMP value is absurd, ignored (-204.0C)
acpi_tz0: _TMP value is absurd, ignored (-204.0C)
acpi_tz0: _TMP value is absurd, ignored (-203.0C)

Found-in: FreeBSD